fix(scheduler): trim the policy in DeviceUsageList.Less
Less dispatches on the whole Policy string, so a padded single value
matched neither the binpack nor the mutex branch and fell through to
spread, which is the opposite order a binpack annotation asks for.
IsValidGPUSchedulerPolicy accepts a padded value and Policy is also
assigned straight from --gpu-scheduler-policy, so both routes reach it.
Compare against a trimmed copy; the chain path already trims each token
in gpuSortKeyChain.

Fold the util cases into scheduler_policy_test.go, which already covered
this function, and drop the two that only asserted the canonical spelling
of a comma separated value, since the chain path normalized those before
this change. Log the policy actually used when an annotation is rejected,
rather than the raw default.
